pyproject.toml
[build-system]
requires = ["setuptools>=65.6.3"]
build-backend = "setuptools.build_meta"

[project]
name = "bleetube-nip05er"
version = "0.1.0"
description = "Internal devops tool for managing inventory of collectors"
readme = "README.md"
authors = [{ name = "blee", email = "blee@bitcoiner.social" }]
license = { file = "LICENSE" }
keywords = ["nostr", "nip05"]
dependencies = [
    'build>=0.8.0',
    'click>=8.1.3',
    'psycopg2>=2.9.5',
    'websockets>=10.4',
    'python-dotenv>=1.0.0',
]
requires-python = ">=3.10.0"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
[project.urls]
Homepage = "https://github.com/bleetube/nip05er"

# https://setuptools.pypa.io/en/latest/userguide/entry_point.html
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
[project.scripts]
nip05er = "bleetube_nip05er.nip05er:cli"